add maxFixedBufferSize materializer setting #19702
